big refactor; separate tests from src
