+++ setup.py
@@ -45,43 +45,43 @@ f90flags.append('-O3')
 f90flags.append('-w')
 
 ext_cape_ml = Extension(name = 'CAPE_CODE_model_lev',
-                        sources = ['xcape/CAPE_CODE_model_lev.pyf',
-                                   'xcape/CAPE_CODE_model_lev.f90'],
+                        sources = ['src/xcape/CAPE_CODE_model_lev.pyf',
+                                   'src/xcape/CAPE_CODE_model_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 ext_cape_pl = Extension(name = 'CAPE_CODE_pressure_lev',
-                        sources = ['xcape/CAPE_CODE_pressure_lev.pyf',
-                                   'xcape/CAPE_CODE_pressure_lev.f90'],
+                        sources = ['src/xcape/CAPE_CODE_pressure_lev.pyf',
+                                   'src/xcape/CAPE_CODE_pressure_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 ext_bunkers_ml = Extension(name = 'Bunkers_model_lev',
-                        sources = ['xcape/Bunkers_model_lev.pyf',
-                                   'xcape/Bunkers_model_lev.f90'],
+                        sources = ['src/xcape/Bunkers_model_lev.pyf',
+                                   'src/xcape/Bunkers_model_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 ext_bunkers_pl = Extension(name = 'Bunkers_pressure_lev',
-                        sources = ['xcape/Bunkers_pressure_lev.pyf',
-                                   'xcape/Bunkers_pressure_lev.f90'],
+                        sources = ['src/xcape/Bunkers_pressure_lev.pyf',
+                                   'src/xcape/Bunkers_pressure_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 ext_srh_ml = Extension(name = 'SREH_model_lev',
-                        sources = ['xcape/SREH_model_lev.pyf',
-                                   'xcape/SREH_model_lev.f90'],
+                        sources = ['src/xcape/SREH_model_lev.pyf',
+                                   'src/xcape/SREH_model_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 ext_srh_pl = Extension(name = 'SREH_pressure_lev',
-                        sources = ['xcape/SREH_pressure_lev.pyf',
-                                   'xcape/SREH_pressure_lev.f90'],
+                        sources = ['src/xcape/SREH_pressure_lev.pyf',
+                                   'src/xcape/SREH_pressure_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 ext_stdh_ml = Extension(name = 'stdheight_2D_model_lev',
-                        sources = ['xcape/stdheight_2D_model_lev.pyf',
-                                   'xcape/stdheight_2D_model_lev.f90'],
+                        sources = ['src/xcape/stdheight_2D_model_lev.pyf',
+                                   'src/xcape/stdheight_2D_model_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 ext_stdh_pl = Extension(name = 'stdheight_2D_pressure_lev',
-                        sources = ['xcape/stdheight_2D_pressure_lev.pyf',
-                                   'xcape/stdheight_2D_pressure_lev.f90'],
+                        sources = ['src/xcape/stdheight_2D_pressure_lev.pyf',
+                                   'src/xcape/stdheight_2D_pressure_lev.f90'],
                              extra_f90_compile_args=f90flags,
                              f2py_options=['--quiet'])
 
@@ -102,7 +102,8 @@ setup(name=DISTNAME,
       install_requires=INSTALL_REQUIRES,
       python_requires=PYTHON_REQUIRES,
       url=URL,
-      packages=find_packages(),
+      package_dir={"": "src"},
+      packages=find_packages("src"),
       ext_package='xcape',
       ext_modules = [ext_cape_ml, ext_cape_pl,
                      ext_bunkers_ml, ext_bunkers_pl,
