+++ src/xcape/core.py
