+++ src/xcape/duck_array_ops.py
